src/ingestor/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-pptx==0.6.21
pdfminer.six==20231228
docx2txt==0.8
pydantic==2.8.2
